finished adding header changes to improve compile speed (at least with MS VC++) and to get rid of some of the clutter that the RTTI macros add. But Where'd the RTTI info go!?!? Have no fear it's still there! It is now located in three .inl files, one for each kit. Each kit will consolidate it's RTTI decls in the following files: FoundationKit:   vcf/include/FoundationKitRTTI.inl GraphicsKit:   vcf/include/GraphicsKitRTTI.inl ApplicationKit:   vcf/include/ApplicationKitRTTI.inl
Please alter these files as neeccessary to add/subtract/alter the various
RTTI info. Each kit is also responsible for registering any of the class
in the ClassRegistry.
A side effect of this is that I had to alter the memory allocation for the
VC6 build for the Application it went from /Zm120 to /Zm150. I find this
pretty lame but I don't know what else to do. Hopefully this will improve
compile times when using the Appkit in other programs.